+++ src/main/java/com/vectorsearch/faiss/swig/LinearTransform.java
@@ -85,15 +85,15 @@ public class LinearTransform extends VectorTransform {
     this(swigfaissJNI.new_LinearTransform__SWIG_3(), true);
   }
 
-  public void apply_noalloc(int n, SWIGTYPE_p_float x, SWIGTYPE_p_float xt) {
+  public void apply_noalloc(long n, SWIGTYPE_p_float x, SWIGTYPE_p_float xt) {
     swigfaissJNI.LinearTransform_apply_noalloc(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(x), SWIGTYPE_p_float.getCPtr(xt));
   }
 
-  public void transform_transpose(int n, SWIGTYPE_p_float y, SWIGTYPE_p_float x) {
+  public void transform_transpose(long n, SWIGTYPE_p_float y, SWIGTYPE_p_float x) {
     swigfaissJNI.LinearTransform_transform_transpose(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(y), SWIGTYPE_p_float.getCPtr(x));
   }
 
-  public void reverse_transform(int n, SWIGTYPE_p_float xt, SWIGTYPE_p_float x) {
+  public void reverse_transform(long n, SWIGTYPE_p_float xt, SWIGTYPE_p_float x) {
     swigfaissJNI.LinearTransform_reverse_transform(swigCPtr, this, n, SWIGTYPE_p_float.getCPtr(xt), SWIGTYPE_p_float.getCPtr(x));
   }
