+++ src/main/java/com/vectorsearch/faiss/swig/MaskedInvertedLists.java
@@ -79,7 +79,7 @@ public class MaskedInvertedLists extends ReadOnlyInvertedLists {
     swigfaissJNI.MaskedInvertedLists_release_ids(swigCPtr, this, list_no, SWIGTYPE_p_long.getCPtr(ids));
   }
 
-  public int get_single_id(long list_no, long offset) {
+  public long get_single_id(long list_no, long offset) {
     return swigfaissJNI.MaskedInvertedLists_get_single_id(swigCPtr, this, list_no, offset);
   }
